api/requirements.txt
Flask_Cors==3.0.9
Flask==1.1.2
psycopg2-binary==2.8.5
